refactor: split oversized slurm/local + cli/commands/jobs modules
Two pure refactors (no behavior change) breaking up >1000-line modules.

slurm/local.py (1029 → 69 lines)
  LocalClient moves into a new slurm/clients/ subpackage at
  clients/local.py. local.py keeps the Slurm legacy wrapper +
  submit_job/retrieve_job/cancel_job convenience functions and
  re-exports LocalClient so existing srunx.slurm.local.LocalClient
  importers keep working. The clients/ namespace is intentionally
  chosen to host clients/ssh.py later when slurm/ssh.py (1887 lines)
  gets the same treatment, making the LocalClient ↔ SlurmSSHAdapter
  sibling relationship explicit.

cli/commands/jobs.py (1282 lines) → cli/commands/jobs/ (package)
  One submodule per Typer subcommand: sbatch / squeue / scancel /
  sinfo / gpus / tail. cli/main.py imports each function from its
  own submodule. The package __init__.py deliberately does NOT
  re-export the command functions — doing so would rebind the
  package attribute so that "import srunx.cli.commands.jobs.sbatch
  as m" returns the function instead of the module, breaking tests
  that monkeypatch module-level attributes. The __init__.py
  docstring documents the rule.

Test patches retargeted where moved code's lookup site changed:
  - srunx.slurm.local.{render_job_script, render_shell_job_script,
    get_job_status} → srunx.slurm.clients.local.<same>
  - srunx.cli.commands.jobs.get_config →
    srunx.cli.commands.jobs.sbatch.get_config

Patches on Slurm.<method> attributes are unchanged — Slurm still
lives in local.py and inherits these methods, so MRO-routed lookups
on Slurm() instances still hit the mock.

Refs #193 (tracking issue for remaining oversized modules).
