+++ src/srunx/slurm/clients/__init__.py
@@ -1,0 +1,14 @@
+"""Concrete SLURM client implementations.
+
+Each module here is a transport-specific implementation of the
+:class:`~srunx.slurm.protocols.Client` Protocol:
+
+- :mod:`srunx.slurm.clients.local` — :class:`LocalClient`, the
+  in-process ``sbatch`` / ``squeue`` / ``scancel`` driver.
+- :mod:`srunx.slurm.clients.ssh` (planned) — the SSH-transport
+  sibling, currently still in :mod:`srunx.slurm.ssh`.
+
+Public consumers should generally import :class:`Slurm` from
+:mod:`srunx.slurm.local` (the legacy wrapper that wires DB recording +
+callbacks) rather than reaching into this subpackage directly.
+"""
